Ensure Astro::Coord::ECI::TLE can raise errors
If the sgp4r method fails, the module attempts to produce an error
message including the time and the object ID.  If object ID is
not set an unnecessary warning is printed.  If the time is missing
then a misleading error is raised saying that the time is missing,
from a routine to which the time is passed!  The actual error
of interest is not reported in that case.  These problems can be
avoided by setting the time before calling sgp4r, and initializing
the object with a dummy object ID.